Delete old SearchSpec
