+++ tests/src/test/resources/kg/search/activity.json
