--- tests/src/test/resources/kg/search/agent1.json
@@ -1,10 +1,0 @@
-{
-  "@context": "https://bbp.neuroshapes.org",
-  "@id": "http://localhost/users/john-doe",
-  "@type": [
-    "Person",
-    "Agent"
-  ],
-  "familyName": "Doe",
-  "givenName": "John"
-}
