--- tests/src/test/resources/kg/search/agent2.json
@@ -1,10 +1,0 @@
-{
-  "@context": "https://bbp.neuroshapes.org",
-  "@id": "http://localhost/users/jane-jones",
-  "@type": [
-    "Person",
-    "Agent"
-  ],
-  "familyName": "Jones",
-  "givenName": "Jane"
-}
