+++ tests/src/test/resources/kg/search/layer-thickness.json
