+++ tests/src/test/resources/kg/search/license.json
