--- tests/src/test/resources/kg/search/org1.json
@@ -1,8 +1,0 @@
-{
-  "@context": "https://bbp.neuroshapes.org",
-  "@id": "http://localhost/orgs/epfl",
-  "@type": [
-    "Organization"
-  ],
-  "name": "EPFL"
-}
