+++ tests/src/test/resources/kg/search/patched-cell.json
