+++ tests/src/test/resources/kg/search/person.json
