+++ tests/src/test/resources/kg/search/protocol.json
