+++ tests/src/test/resources/kg/search/trace.json
